+++ src/YCHR/Internal/Repl.hs
@@ -776,7 +776,7 @@ renderInfoEntry = \case
   IEFunction f kind ->
     prettyQualifiedName f.name
       ++ "\n"
-      ++ prettyFunctionDecl f.name f.arity f.signatures f.requiring kind
+      ++ prettyFunctionDecl f.name f.arity f.signatures f.requiring f.refining kind
       ++ "\n"
   IEType td ->
     let qn = case td.name of