+++ test/YCHR/RenameTest.hs
@@ -1390,6 +1390,7 @@ importListTests =
                                     False
                                     DKFunction
                                     Nothing
+                                    Nothing
                                 ]
                             )
                         )